refactor(web): changed name from TooltipTest to Tooltip
