+++ packages/web/src/components/components.ts
@@ -18,13 +18,12 @@ export * from './RadioInput'
 export * from './Select'
 export * from './FileInput'
 export * from './Slider'
-export * from './Tooltip'
 export * from './List'
 export * from './Link'
 export * from './LoadingOverlay'
 export * from './InputBase'
 export * from './Switch'
 export * from './NumberIncrement'
-export * from './TooltipTest'
+export * from './Tooltip'
 
 export * from './defaultStyles'
