+++ packages/web/src/components/defaultStyles.ts
@@ -14,7 +14,6 @@ import { OverlayPresets } from './Overlay/styles'
 import { TextInputPresets } from './TextInput/styles'
 import { RadioInputPresets } from './RadioInput/styles'
 import { SelectPresets } from './Select/styles'
-import { TooltipPresets } from './Tooltip/styles'
 import { ListPresets } from './List/styles'
 import { LinkPresets } from './Link/styles'
 import { SliderPresets } from './Slider/styles'
@@ -22,7 +21,7 @@ import { LoadingOverlayPresets } from './LoadingOverlay/styles'
 import { InputBasePresets } from './InputBase'
 import { SwitchPresets } from './Switch/styles'
 import { NumberIncrementPresets } from './NumberIncrement/styles'
-import { TooltipTestPresets } from './TooltipTest/styles'
+import { TooltipPresets } from './Tooltip/styles'
 
 export const defaultStyles = {
   View: ViewPresets,
@@ -42,12 +41,11 @@ export const defaultStyles = {
   TextInput: TextInputPresets,
   RadioInput: RadioInputPresets,
   Select: SelectPresets,
-  Tooltip: TooltipPresets,
   List: ListPresets,
   Slider: SliderPresets,
   LoadingOverlay: LoadingOverlayPresets,
   InputBase: InputBasePresets,
   Switch: SwitchPresets,
   NumberIncrement: NumberIncrementPresets,
-  TooltipTest: TooltipTestPresets,
+  Tooltip: TooltipPresets,
 }
